Регистрация ч.2

+++ database/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Time, Boolean, ForeignKey
+from sqlalchemy import Column, Integer, String, Time, Boolean, ForeignKey, BigInteger
 from sqlalchemy.orm import relationship, backref
 from sqlalchemy.ext.declarative import declarative_base
 
@@ -10,7 +10,7 @@ class User(Base):
     __tablename__ = 'users'
 
     id = Column(Integer, primary_key=True)
-    chat_id = Column(Integer, nullable=False)
+    chat_id = Column(BigInteger, nullable=False)
     gruppa = Column(ForeignKey('groups.id'), default=None)
     formob = Column(String(1), default=None)
     kyrs = Column(Integer(), default=None)
